refactor: remove duplicates, refactor to reduce complexity budget
