+++ pkg/integrations/cloudflare/cloudflare.go
@@ -179,10 +179,6 @@ func (c *Cloudflare) Actions() []core.Action {
 		&UpdatePool{},
 		&GetPool{},
 		&DeletePool{},
-		&CreatePool{},
-		&UpdatePool{},
-		&GetPool{},
-		&DeletePool{},
 		&CreateLoadBalancer{},
 		&GetLoadBalancer{},
 		&UpdateLoadBalancer{},
